Implement `drawTriangleRasterization`, fix minor build issues
