+++ src/Primitive.h
@@ -1,0 +1,7 @@
+#pragma once
+
+typedef enum
+{
+    PRIMITIVE_TRIANGLES
+} Primitive;
+
